+++ packages/web3-eth/src/utils/transaction_builder.ts
@@ -48,6 +48,7 @@ import { bytesToHex, format } from 'web3-utils';
 import { NUMBER_DATA_FORMAT } from '../constants.js';
 // eslint-disable-next-line import/no-cycle
 import { getChainId, getTransactionCount, estimateGas } from '../rpc_method_wrappers.js';
+// eslint-disable-next-line import/no-cycle
 import { detectTransactionType } from './detect_transaction_type.js';
 import { transactionSchema } from '../schemas.js';
 import { InternalTransaction } from '../types.js';
@@ -109,11 +110,11 @@ export const getTransactionNonce = async <ReturnFormat extends DataFormat>(
 	return getTransactionCount(web3Context, address, web3Context.defaultBlock, returnFormat);
 };
 
-export const getTransactionType = (
+export const getTransactionType = async (
 	transaction: FormatType<Transaction, typeof ETH_DATA_FORMAT>,
 	web3Context: Web3Context<EthExecutionAPI>,
 ) => {
-	const inferredType = detectTransactionType(transaction, web3Context);
+	const inferredType = await detectTransactionType(transaction, web3Context);
 	if (!isNullish(inferredType)) return inferredType;
 	if (!isNullish(web3Context.defaultTransactionType))
 		return format({ format: 'uint' }, web3Context.defaultTransactionType, ETH_DATA_FORMAT);
@@ -214,7 +215,7 @@ export async function defaultTransactionBuilder<ReturnType = Transaction>(option
 		populatedTransaction.gasLimit = populatedTransaction.gas;
 	}
 
-	populatedTransaction.type = getTransactionType(populatedTransaction, options.web3Context);
+	populatedTransaction.type = await getTransactionType(populatedTransaction, options.web3Context);
 	if (
 		isNullish(populatedTransaction.accessList) &&
 		(populatedTransaction.type === '0x1' || populatedTransaction.type === '0x2')