+++ packages/web3-eth/test/integration/defaults.test.ts
@@ -761,7 +761,7 @@ describe('defaults', () => {
 			});
 			expect(eth2.defaultCommon).toBe(common);
 		});
-		it('defaultTransactionType', () => {
+		it('defaultTransactionType', async () => {
 			// default
 			expect(web3Eth.defaultTransactionType).toBe('0x2');
 			// after set
@@ -778,7 +778,7 @@ describe('defaults', () => {
 			});
 			expect(eth2.defaultTransactionType).toBe('0x4444');
 
-			const res = getTransactionType(
+			const res = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -798,7 +798,7 @@ describe('defaults', () => {
 			// tx.maxPriorityFeePerGas !== undefined ||
 			// tx.hardfork === 'london' ||
 			// tx.common?.hardfork === 'london'
-			const maxFeePerGasOverride = getTransactionType(
+			const maxFeePerGasOverride = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -813,7 +813,7 @@ describe('defaults', () => {
 				eth2,
 			);
 			expect(maxFeePerGasOverride).toBe('0x2');
-			const maxPriorityFeePerGasOverride = getTransactionType(
+			const maxPriorityFeePerGasOverride = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -828,7 +828,7 @@ describe('defaults', () => {
 				eth2,
 			);
 			expect(maxPriorityFeePerGasOverride).toBe('0x2');
-			const hardforkOverride = getTransactionType(
+			const hardforkOverride = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -843,7 +843,7 @@ describe('defaults', () => {
 				eth2,
 			);
 			expect(hardforkOverride).toBe('0x2');
-			const commonOverride = getTransactionType(
+			const commonOverride = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -865,7 +865,7 @@ describe('defaults', () => {
 			// override to 0x1 if:
 			// tx.accessList !== undefined || tx.hardfork === 'berlin' || tx.common?.hardfork === 'berlin'
 
-			const accessListOverride = getTransactionType(
+			const accessListOverride = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -886,7 +886,7 @@ describe('defaults', () => {
 			);
 			expect(accessListOverride).toBe('0x1');
 
-			const hardforkBerlinOverride = getTransactionType(
+			const hardforkBerlinOverride = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -902,7 +902,7 @@ describe('defaults', () => {
 			);
 			expect(hardforkBerlinOverride).toBe('0x0');
 
-			const commonBerlinOverride = getTransactionType(
+			const commonBerlinOverride = await getTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',
@@ -1013,7 +1013,7 @@ describe('defaults', () => {
 			});
 			expect(newBuilderMock).toHaveBeenCalled();
 		});
-		it('transactionTypeParser', () => {
+		it('transactionTypeParser', async () => {
 			// default
 			expect(web3Eth.transactionTypeParser).toBeUndefined();
 
@@ -1031,7 +1031,7 @@ describe('defaults', () => {
 				},
 			});
 			expect(eth2.transactionTypeParser).toBe(newParserMock);
-			detectTransactionType(
+			await detectTransactionType(
 				{
 					from: '0xEB014f8c8B418Db6b45774c326A0E64C78914dC0',
 					to: '0x3535353535353535353535353535353535353535',